refactor(billing): move pay-as-you-go into the current plan card
Pay-as-you-go now renders at the bottom of the Current Plan card instead
of a standalone box. Spend caps show as read-only labels; enabling and
editing them happen in a modal.
